Bookstore Project - Added Review Management Module

+++ BookStoreProject/BookStoreWebsite/src/com/bookstore/service/CustomerServices.java
@@ -105,6 +105,13 @@ public class CustomerServices {
 			CommonUtility.showMessageBackend(request, response, message);
 			return;
 		}
+		
+		if (customerDAO.hasReviews(customerId)) {
+			String message = "Could not delete customer with ID " + customerId
+					+ " because it has one or more reviews";
+			CommonUtility.showMessageBackend(request, response, message);
+			return;
+		}
 
 		customerDAO.delete(customerId);
